auth-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-prometheus
gunicorn
